backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & Auth
supabase==2.0.2
psycopg2-binary==2.9.9

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# AI & ML
openai==1.3.7
anthropic==0.7.8
groq==0.4.1

# Utils
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0 